requirements.txt
datamol==0.12.4
fcd-torch==1.0.7
matplotlib==3.8.4
molfeat==0.10.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
tabulate==0.9.0
tokenizers==0.15.2
tqdm==4.66.2
transformers==4.39.3
xgboost==2.0.3
# OPTIONAL
# submitit==1.5.1   # for submitting jobs to slurm